refactor(group): separate manifest helper utilities
